+++ filters/common/cookiefs.c
@@ -144,7 +144,8 @@ cosign_cookie_valid( cosign_host_config *cfg, char *cookie, struct sinfo *si,
 
     /* store local copy of scookie (service cookie) */
     if ( snprintf( tmppath, sizeof( tmppath ), "%s/%x%x.%i", cfg->filterdb,
-	    tv.tv_sec, tv.tv_usec, (int)getpid()) >= sizeof( tmppath )) {
+	    (int)tv.tv_sec, (int)tv.tv_usec, (int)getpid()) >=
+	    sizeof( tmppath )) {
 	cosign_log( APLOG_ERR, s, "mod_cosign: cosign_cookie_valid: "
 		"tmppath too long" );
 	return( -1 );
